Migrate package to opensearch
To start getting our various test environments going with
opensearch it would be useful to have a docker image with at
least some of our plugins available. The docker image uses the
plugins .deb to ensure it has the same set of plugins we see in
production without duplicating. As such, migrate this debian
package over to opensearch to be used there.

Opensearch changes signature verification to .sig files instead
of .asc files. Appears to work as expected by changing the file
extension.

This contains all plugins that were available for elasticsearch,
nothing should be lost in the shift to opensearch.

Bug: T372769
Change-Id: I50a8f0ea5085b9f8b0c3627932f198aef0ed00ea