Add sudachi analyzer for japanese
For the most part follows our standard method of adding plugins,
but this plugin requires an additional dictionary installed
outside of the plugin folder. To support this an additional
script was added. The script could probably be generalized, but
as we only have a single use-case today it seems premature.

The 1.3.20 version of the plugin comes from people.wikimedia.org
as we had to run a custom backport, but once we upgrade to
opensearch 2.x we can switch to an upstream package.

Bug: T386868
Change-Id: I3296d6df0632737c8be3db6a7b917e0a8c1d64b3